Add all sage builtin functions (#542)
Issue #537
Added all built-in sage properties and invariants, up to Sage 8.2.
Added doctests to automatically ensure we check for new/changed functions in future versions of Sage.
Split sage_invariants into efficient and intractable lists.
Removed any now-duplicate functions we created in favor of the built-in functions, except for is_two_connected because we prefer that to is_biconnected. Note, though, that we redefined is_two_connected to raise ValueError if order() < 2, due to debate introduced by the definition of is_biconnected.

Squashed merge to clean up many typo fixes.